add: double to float, mean loss as end epoch's callback parameter

+++ layer.cpp
@@ -3,27 +3,27 @@
 
 DenseLayer::DenseLayer(int input_size, int output_size) {
     weights = xt::random::rand({output_size, input_size}, -1.0f, 1.0f);
-    biases = xt::zeros<double>({output_size}); 
+    biases = xt::zeros<float>({output_size}); 
 }
 
-xt::xarray<double> DenseLayer::forward(const xt::xarray<double>& inputs) {
+xt::xarray<float> DenseLayer::forward(const xt::xarray<float>& inputs) {
     this->inputs = inputs;
     return xt::linalg::dot(inputs, xt::transpose(weights)) + biases; 
 }
 
-xt::xarray<double> DenseLayer::backward(const xt::xarray<double>& upstream_gradient, double lr) {
+xt::xarray<float> DenseLayer::backward(const xt::xarray<float>& upstream_gradient, float lr) {
     auto weights_gradient = xt::linalg::outer(inputs, upstream_gradient);
     weights -= lr * xt::transpose(weights_gradient);
     biases -= lr * upstream_gradient;
     return xt::linalg::dot(upstream_gradient, weights);
 }
 
-double SigmoidLayer::activation_function(double weighted_sum) {
+float SigmoidLayer::activation_function(float weighted_sum) {
     return 1.0 / (1.0 + std::exp(-weighted_sum));
 }
 
-xt::xarray<double> SigmoidLayer::forward(const xt::xarray<double>& inputs) {
-    auto vecf = xt::vectorize([this](double x) {
+xt::xarray<float> SigmoidLayer::forward(const xt::xarray<float>& inputs) {
+    auto vecf = xt::vectorize([this](float x) {
         return this->activation_function(x);
     });
 
@@ -31,17 +31,17 @@ xt::xarray<double> SigmoidLayer::forward(const xt::xarray<double>& inputs) {
     return last_output;
 }
 
-xt::xarray<double> SigmoidLayer::backward(const xt::xarray<double>& upstream_gradient, double lr) {
+xt::xarray<float> SigmoidLayer::backward(const xt::xarray<float>& upstream_gradient, float lr) {
     return upstream_gradient * last_output * (1 - last_output);
 }
 
 
-double ReLULayer::activation_function(double weighted_sum) {
-    return std::max((double)0.0f, weighted_sum);
+float ReLULayer::activation_function(float weighted_sum) {
+    return std::max((float)0.0f, weighted_sum);
 }
 
-xt::xarray<double> ReLULayer::forward(const xt::xarray<double>& inputs) {
-    auto vecf = xt::vectorize([this](double x) {
+xt::xarray<float> ReLULayer::forward(const xt::xarray<float>& inputs) {
+    auto vecf = xt::vectorize([this](float x) {
         return this->activation_function(x);
     });
 
@@ -50,6 +50,6 @@ xt::xarray<double> ReLULayer::forward(const xt::xarray<double>& inputs) {
     return last_output;
 }
 
-xt::xarray<double> ReLULayer::backward(const xt::xarray<double>& upstream_gradient, double lr) {
-    return upstream_gradient * xt::cast<double>(this->inputs > 0);
+xt::xarray<float> ReLULayer::backward(const xt::xarray<float>& upstream_gradient, float lr) {
+    return upstream_gradient * xt::cast<float>(this->inputs > 0);
 }
